# Conflicts:
#	conf/messages.en

+++ app/viewmodels/checkAnswers/SupplementaryUnitSummary.scala
@@ -66,7 +66,7 @@ object SupplementaryUnitSummary {
       )
     }
 
-  def row(suppValue: Option[Double], measureValue: Option[String], recordId: String)(implicit
+  def row(suppValue: Option[BigDecimal], measureValue: Option[String], recordId: String)(implicit
     messages: Messages
   ): Option[SummaryListRow] =
     for {
